shared fixtures for all tests
