+++ tests/conftest.py
@@ -2,6 +2,7 @@ import contextlib
 import json
 import logging
 import os
+import shutil
 import tempfile
 from pathlib import Path
 from typing import Any
@@ -11,10 +12,14 @@ from typing import List
 from unittest.mock import patch
 
 import pytest
+from expected_download import _get_files_in_directory
 
+from ytdl_sub.config.config_file import ConfigFile
+from ytdl_sub.subscriptions.subscription_download import SubscriptionDownload
 from ytdl_sub.utils.file_handler import FileHandler
 from ytdl_sub.utils.logger import Logger
 from ytdl_sub.utils.logger import LoggerLevels
+from ytdl_sub.utils.yaml import load_yaml
 
 
 @pytest.fixture(autouse=True)
@@ -28,9 +33,31 @@ def cleanup_debug_file():
 
 
 @pytest.fixture
-def working_directory() -> Path:
+def working_directory() -> str:
+    """
+    Any time the working directory is used, ensure no files remain on cleaning it up
+    """
+    logger = Logger.get("test")
+
     with tempfile.TemporaryDirectory() as temp_dir:
-        yield temp_dir
+
+        def _assert_working_directory_empty(self, is_error: bool = False):
+            files = [str(file_path) for file_path in _get_files_in_directory(temp_dir)]
+            num_files = len(files)
+            if os.path.isdir(temp_dir):
+                shutil.rmtree(temp_dir)
+
+            if not is_error:
+                if num_files > 0:
+                    logger.error("left-over files in working dir:\n%s", "\n".join(files))
+                assert num_files == 0
+
+        with patch.object(
+            SubscriptionDownload,
+            "_delete_working_directory",
+            new=_assert_working_directory_empty,
+        ):
+            yield temp_dir
 
 
 @pytest.fixture()
@@ -106,3 +133,38 @@ def preset_dict_to_subscription_yaml_generator() -> Callable:
             FileHandler.delete(tmp_file.name)
 
     return _preset_dict_to_subscription_yaml_generator
+
+
+###################################################################################################
+# Example config fixtures
+
+
+def _load_config(config_path: Path, working_directory: Path) -> ConfigFile:
+    config_dict = load_yaml(file_path=config_path)
+    config_dict["configuration"]["working_directory"] = working_directory
+
+    return ConfigFile.from_dict(config_dict)
+
+
+@pytest.fixture()
+def music_video_config_path() -> Path:
+    return Path("examples/music_videos_config.yaml")
+
+
+@pytest.fixture()
+def music_video_config(music_video_config_path, working_directory) -> ConfigFile:
+    return _load_config(music_video_config_path, working_directory)
+
+
+@pytest.fixture()
+def channel_as_tv_show_config(working_directory) -> ConfigFile:
+    return _load_config(
+        config_path=Path("examples/tv_show_config.yaml"), working_directory=working_directory
+    )
+
+
+@pytest.fixture()
+def music_audio_config(working_directory) -> ConfigFile:
+    return _load_config(
+        config_path=Path("examples/music_audio_config.yaml"), working_directory=working_directory
+    )
